requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
anthropic
groq
google-genai